Style: move structs to other location
